Use double for destiny m_targetDistance instead of uint32 (#295)
* use double for destiny m_targetDistance instead of uint32; leverage halting after warp stop for slightly better stability

* add a warp sysbubble at each tick during warp; various formatting and logging improvements and fixes

* code cleanup/add comments

+++ src/eve-server/system/DestinyManager.h
@@ -296,7 +296,7 @@ protected:
     float m_maxOrbitSpeedFraction;      //fuzzy logic - ship's max speed based on orbit data
 
     uint32 m_followDistance;            //in m
-    uint32 m_targetDistance;            //in m
+    double m_targetDistance;            //in m
     double m_moveTime;                  //in ms       - time when speed change started.  used to calculate m_timeFraction
     double m_callTime;                  //in ms       - time client call was processed.  this is to coordinate tic calculations
